pyproject.toml
[project]
name = "repl-nix-youtubearticlegenerator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.151.0",
    "google-generativeai>=0.8.3",
    "postgrest-py>=0.10.6",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.36",
    "streamlit>=1.40.0",
    "youtube-transcript-api>=0.6.2",
]
